requirements.txt
streamlit
fpdf
qrcode[pil]
python-dotenv